Made isSqrt accept long long int

+++ Maths/Divisors/isSqrt/isSqrt.c
@@ -2,7 +2,7 @@
 #include "../sqrt/sqrt.c"
 
 //Gives 1 if n is a perfect square.
-int isSqrt(int n){
+int isSqrt(long long int n){
     int i = sqrtN(n);
     if(n==i*i)
         return 1;
